requirements.txt
streamlit
opencv-python-headless
numpy
scikit-learn
joblib
tensorflow
pillow
